+++ modal.js
@@ -11,6 +11,28 @@
 	'use strict';
 
 	/*
+	 * Polyfill CustomEvent
+	 */
+	var CustomEvent = function (event, params) {
+		var evt = document.createEvent('CustomEvent');
+
+		params = params || {
+			bubbles: false,
+			cancelable: false,
+			detail: undefined
+		};
+
+		evt.initCustomEvent(event, params.bubbles, params.cancelable, params.detail );
+
+		return evt;
+	};
+
+	if (!window.CustomEvent) {
+		CustomEvent.prototype = window.Event.prototype;
+		window.CustomEvent = CustomEvent;
+	}
+
+	/*
 	 * Storage for functions and attributes
 	 */
 	var modal = {
@@ -53,13 +75,7 @@
 		 * @param modal {string} id of modal that the event is triggered on
 		 */
 		trigger: function (event, modal) {
-			var eventTrigger;
-
-			if (!window.CustomEvent) {
-				return;
-			}
-
-			eventTrigger = new CustomEvent(event, {
+			var eventTrigger = new CustomEvent(event,{
 				detail: {
 					'modal': modal
 				}